// vibeui/src/app/api/model/itemSummary.ts
/**
 * CRISTAL-iSE Rest API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ViewpointSummary } from './viewpointSummary';
import { PropertySummary } from './propertySummary';
import { CollectionSummary } from './collectionSummary';


export interface ItemSummary { 
    uuid: string;
    name: string;
    domainPaths: Array<string>;
    hasMasterOutcome: boolean;
    isAgent: boolean;
    type?: string;
    master?: string;
    properties?: Array<PropertySummary>;
    viewpoints?: Array<ViewpointSummary>;
    collections?: Array<CollectionSummary>;
    workflow: string;
    history: string;
    outcome: string;
    attachment: string;
    job: string;
    roles?: string;
}